Update Julia packages

+++ app/src/realtime-controller/pid.jl
@@ -11,7 +11,7 @@ struct PIDController <: RTController
     pid::DiscretePID
 end
 
-function PIDController(resolution, Kp, Ti, Td)
+function PIDController(resolution, Kp::Float64, Ti::Float64, Td::Float64)
     pid = DiscretePID(; K = Kp, Ts = resolution / Second(1), Ti = Ti, Td = Td)
     return PIDController(resolution, Kp, Ti, Td, pid)
 end
